Removing PPLAY and merging directory organization from develop
