--- PPlay/__init__.py
@@ -1,12 +1,0 @@
-__all__ =   [
-            "animation",
-            "collision",
-            "gameimage",
-            "gameobject",
-            "keyboard",
-            #"mouse",
-            "point",
-            #"sound",
-            "sprite",
-            "window"
-            ]
