--- PPlay/point.py
@@ -1,7 +1,0 @@
-# coding= utf-8
-
-"""A Point class"""
-class Point():
-    def __init__(self, x, y):
-        self.x = x
-        self.y = y
