+++ src/Application/observer.py
@@ -1,11 +1,10 @@
-from PPlay.window import *
-from scenes.menu import *
+from Lib.window import *
+from Scenes.menu import *
 
-janela = Window(1080, 600)
-janela.set_background_color((0, 0, 0))
-janela.set_title("Xadrez")
-mouse = janela.get_mouse()
-keyboard = janela.get_keyboard()
+window = Window(1080, 600)
+window.set_background_color((0, 0, 0))
+window.set_title("Xadrez")
+mouse = window.get_mouse()
 
 dif = 0
 actual_scene = 0
@@ -13,7 +12,7 @@ actual_scene = 0
 def change_scene(scene):
     global actual_scene,dif
     if scene == 1:
-        actual_scene = Menu(janela, mouse, keyboard)
+        actual_scene = Menu(window, mouse)
         actual_scene.start()
     else:
         quit()
@@ -23,7 +22,7 @@ def draw():
 
 def update():
     global dif
-    janela.update()
+    window.update()
     actual_scene.update()
     if actual_scene.change != 0:
         try:
